requirements.txt
pytest==7.4.2
pytest-cov==4.1.0
pylint==3.0.1
black==23.10.0
fire==0.5.0
fastapi==0.104.0
uvicorn==0.23.2
httpx==0.25.0
icmplib==3.0.4
netmiko==4.2.0
pymongo==4.5.0
chromadb==0.4.14
python-decouple==3.8
langchain==0.0.308
python-multipart==0.0.6
motor==3.3.1